// src/lib/types/PersonalData.ts
export type PersonalDataModel = {
  personalDataId: string;
  employeeNo: string;
  biometricNo: string;
  userRole: string;
  employeeId: string;
  surname: string;
  firstname: string;
  middlename: string;
  extname: string;
  dob: string;
  pob: string;
  sex_id: number;
  civilStatus_id: number;
  height: string;
  weight: string;
  bloodType: string;
  gsisId: string;
  pagibigId: string;
  philhealthNo: string;
  sssNo: string;
  tinNo: string;
  agencyEmpNo: string;
  citizenship: string;
  resAddress: string;
  resZip: string;
  permAddress: string;
  permZip: string;
  telNo: string;
  mobileNo: string;
  email: string;
  employeePicture: File | string | null;
  employeeSignature: File | string | null;
  spouseSurname?: string;
  spouseFirstname?: string;
  spouseMiddlename?: string;
  spouseOccupation?: string;
  spouseEmployer?: string;
  spouseBusinessAddress?: string;
  spouseTelNo?: string;
  fatherSurname?: string;
  fatherFirstname?: string;
  fatherMiddlename?: string;
  motherSurname?: string;
  motherFirstname?: string;
  motherMiddlename?: string;
  govIdNumber: string;
  govIdType: string;
  govIdDate: string;
  govIdPlace: string;
  skillOrHobby?: string;
  distinction?: string;
  association?: string;
  q34a?: string;
  q34b?: string;
  q35a?: string;
  q35b?: string;
  q36?: string;
  q37a?: string;
  q37b?: string;
  q37c?: string;
  q38?: string;
  q39a?: string;
  q39b?: string;
  q39c?: string;
  q34aDetails?: string;
  q34bDetails?: string;
  q35aDetails?: string;
  q35bDetails?: string;
  q35bDateFiled?: string;
  q35bStatus?: string;
  q36Details?: string;
  q37aDetails?: string;
  q37bDetails?: string;
  q37cDetails?: string;
  q38Details?: string;
  q39aDetails?: string;
  q39bDetails?: string;
  q39cDetails?: string;
  q42?: boolean;
};